Add an integration test of excluding dependencies of roots.
