+++ testprojects/src/java/org/pantsbuild/testproject/phrases/TrustyCompanion.java
@@ -1,0 +1,5 @@
+package org.pantsbuild.testproject.phrases;
+
+public class TrustyCompanion {
+  public static String VALUE = "trusty companion";
+}
